+++ ibdl-extractors/src/websites/danbooru/mod.rs
@@ -66,7 +66,7 @@ impl Extractor for DanbooruExtractor {
         }
     }
 
-    async fn search(&mut self, page: usize) -> Result<PostQueue, ExtractorError> {
+    async fn search(&mut self, page: u16) -> Result<PostQueue, ExtractorError> {
         let mut posts = Self::get_post_list(self, page).await?;
 
         if posts.is_empty() {
@@ -87,8 +87,8 @@ impl Extractor for DanbooruExtractor {
 
     async fn full_search(
         &mut self,
-        start_page: Option<usize>,
-        limit: Option<usize>,
+        start_page: Option<u16>,
+        limit: Option<u16>,
     ) -> Result<PostQueue, ExtractorError> {
         let blacklist = BlacklistFilter::init(
             ImageBoards::Danbooru,
@@ -99,7 +99,7 @@ impl Extractor for DanbooruExtractor {
         .await?;
 
         let mut fvec = if let Some(size) = limit {
-            Vec::with_capacity(size)
+            Vec::with_capacity(size as usize)
         } else {
             Vec::new()
         };
@@ -134,7 +134,7 @@ impl Extractor for DanbooruExtractor {
             fvec.extend(list);
 
             if let Some(num) = limit {
-                if fvec.len() >= num {
+                if fvec.len() >= num as usize {
                     break;
                 }
             }
@@ -161,7 +161,7 @@ impl Extractor for DanbooruExtractor {
         Ok(fin)
     }
 
-    async fn get_post_list(&self, page: usize) -> Result<Vec<Post>, ExtractorError> {
+    async fn get_post_list(&self, page: u16) -> Result<Vec<Post>, ExtractorError> {
         let url = format!(
             "{}?tags={}",
             ImageBoards::Danbooru.post_url(),
